data: represent scalars as `f32`
Summary:
In-memory and wire representations of scalars are now 32-bit rather than
64-bit floats. We no longer support reading `DT_DOUBLE` tensors, which
is fine because TensorBoard has never written such tensors to event
files. This is mostly a no-op on the Python side, except that Python
`float`-to-`float` conversions that go through the data provider layer
are now lossy.

Per a comment of @nfelt on #4352.

Test Plan:
All the Rust changes follow from the types. Tests suffice, as usual.

wchargin-branch: data-scalars-f32
wchargin-source: c3f4dacebda6f33381389643458d450aaf5c8926

+++ tensorboard/data/server/logdir.rs
@@ -341,7 +341,7 @@ mod tests {
                         .collect();
                     (run, values)
                 })
-                .collect::<HashMap<&Run, Vec<f64>>>(),
+                .collect::<HashMap<&Run, Vec<f32>>>(),
             expected_data
         );
